% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoRDM.R
\name{build_fordm_table}
\alias{build_fordm_table}
\title{Build FoRDM Table}
\usage{
build_fordm_table(data, management, sow, time)
}
\arguments{
\item{data}{A data.frame containing the input data.}

\item{management}{The name of the management column.}

\item{sow}{The name of the state-of-the-worlds (SOW) column.}

\item{time}{The name of the time column.}
}
\value{
A list with the processed data for further use in the FoRDM analysis, including the input data, mapping for identification of columns and objective columns.
}
\description{
Transfers the provided data table into the format for FoRDM analysis. The columns that represent management, sow (state-of-the-world, scenarios), and time have to be defined. All other columns are treated as objectives.
}
