% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoRDM-package.R
\docType{package}
\encoding{UTF-8}
\name{FoRDM}
\alias{FoRDM-package}
\alias{FoRDM}
\title{Forest related multi-objective Robust Decision Making (FoRDM)}
\description{
This package provides a toolkit for Forest research related many-objective Robust Decision Making (FoRDM)
analysis under uncertainty. It includes robustness measure calculation of regret-based (Type II and cVaR) and satisficing-based as well as robustness
trade-off analysis to support decision making in forest ecosystems under deep uncertainty.
}
\details{
Main features:
\itemize{
\item Build FoRDM input tables from simulation outputs.
\item Compute robustness measures: Regret nd Satisficing.
\item Visualize output in 2D-, 3D- or Parallel Coordinate plots.
\item Perform robustness trade-off analysis.
}
}
\seealso{
browseVignettes(package = "FoRDM")
}
\author{
Marc Djahangard \email{marc.djahangard@ife.uni-freiburg.de} (author, maintainer)

Rasoul Yousefpour (author)
}
