% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoRDM.R
\name{visualize_fordm_2d}
\alias{visualize_fordm_2d}
\title{Visualize 2D Pareto Front}
\usage{
visualize_fordm_2d(analysis_output, x, y, fordm_method)
}
\arguments{
\item{analysis_output}{Output list from FoRDM_analysis_regret() or FoRDM_analysis_satisficing().}

\item{x}{Name of the objective for the x-axis (string).}

\item{y}{Name of the objective for the y-axis (string).}

\item{fordm_method}{Either "regret" or "satisficing".}
}
\value{
A ggplot2 object showing the 2D Pareto front for the selected objectives.
}
\description{
Plots a 2D plot of the Pareto front of management alternatives from FoRDM_analysis output.
}
