% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoRDM.R
\name{visualize_fordm_3d}
\alias{visualize_fordm_3d}
\title{Visualize 3D Pareto Front}
\usage{
visualize_fordm_3d(analysis_output, x, y, z, fordm_method)
}
\arguments{
\item{analysis_output}{Output from FoRDM_analysis_regret or FoRDM_analysis_satisficing.}

\item{x}{Name of the objective for the x-axis (string).}

\item{y}{Name of the objective for the y-axis (string).}

\item{z}{Name of the objective for the z-axis (string).}

\item{fordm_method}{Either "regret" or "satisficing".}
}
\value{
A plotly object showing the 3D Pareto front for the selected objectives.
}
\description{
Plots a 3D plot of the Pareto front of management alternatives from FoRDM_analysis output.
}
