% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoRDM.R
\name{visualize_fordm_parcoord}
\alias{visualize_fordm_parcoord}
\title{Visualize a Parallel Coordinates plot of the Pareto Front for FoRDM Analysis Results}
\usage{
visualize_fordm_parcoord(analysis_output, fordm_method)
}
\arguments{
\item{analysis_output}{Output from fordm_analysis_regret() or fordm_analysis_satisficing().}

\item{fordm_method}{Either "regret" or "satisficing".}
}
\value{
A parallel coordinates plot object.
}
\description{
Creates a parallel coordinates plot showing the Pareto front from FoRDM analysis.
}
